requirements.txt
streamlit
xai-sdk
Pillow